UnitfulMoles and functional trait ontology Have included the package UnitfulMoles to obtain molar mass of water vapour and air - this means that dryair and wetair are now explicitly computing this based on the fractional amounts of O2, CO2 and N2 in the air and they thus need these fractions as inputs.
Also, We’re putting together a paper on how to build a functional trait database for biophysical models and how to construct the trait ontology, which should link to the type system in Julia. At the moment I’m thinking of the following hierarchy, based on the idea that all traits can be classed as either morphological, physiological or behavioural and, following the Functional Ecology paper on functional traits, as ‘model’ (relating to the model structure, e.g. ectotherm/endotherm), ‘parameter’ (a model parameter), ‘threshold’ (a state variable threshold where something happens), or ‘estimation’ (used to obtain the latter three):

AbstractFunctionalTrait
	AbstractMorphologyTraits
		AbstractMorphoModel
		AbstractMorphoParameters
		AbstractMorphoThresholds
	AbstractPhysiologyTraits
		AbstractPhysioModel
		AbstractPhysioParameters
		AbstractPhysioThresholds
	AbstractBehaviorTraits
		AbstractBehavModel
		AbstractBehavParameters
		AbstractBehavThresholds